experimentation/pyproject.toml
[project]
name = "experimentation"
version = "0.1.0"
description = "Experimentation of AI solutions for solving Github Issues."
requires-python = ">=3.9,<3.10"
authors = [
    { name = "Bruno Scaglione", email = "bruno.c.scaglione@gmail.com" }
]
dependencies = [
    "pydantic",
    "pyyaml",
    "pandas",
    "datasets",
    "python-dotenv",
    "vertexai",
    "litellm>=1.57.2",
    "rich>=13.9.4",
    "openai>=1.59.5",
]

[[tool.uv.index]]
url = "https://test.pypi.org/simple/"
default = true

[[tool.uv.index]]
url = "https://pypi.org/simple/"


